requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.22
asyncpg==0.30.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.5.1
jinja2==3.1.2
pdfplumber
python-docx
numpy
sentence-transformers
requests
aiohttp